player info refactoring
